エラーハンドリング

+++ app/app.js
@@ -5,6 +5,7 @@ const path = require('path')
 const bodyParser = require('body-parser')
 //basic認証
 var basicAuth = require('basic-auth-connect');
+const { send } = require('process')
 
 const dbPath = "app/db/database.sqlite3"
 
@@ -41,9 +42,14 @@ app.get('/api/v1/users/:id',(req, res) => {
     const db = new sqlite3.Database(dbPath)
     const id = req.params.id
 
-    db.get(`SELECT * FROM users WHERE id=${id}`, (err,rows) => {
-        //resはjsonというメソッドを持っていて引数で指定したjsonをそのまま返す。
-        res.json(rows)
+    db.get(`SELECT * FROM users WHERE id=${id}`, (err,row) => {
+        if(!row) {
+            // res.status(404).send({error:"Not Found!"})
+            res.status(404).send({error:"Not Found!"})
+        } else {
+            //resはjsonというメソッドを持っていて引数で指定したjsonをそのまま返す。
+            res.status(200).json(row)
+        }
     })
     db.close()
 })
@@ -61,14 +67,12 @@ app.get('/api/v1/search',(req, res) => {
     db.close()
 })
 
-const run = async (sql,db,res,message) => {
+const run = async (sql,db) => {
     return new Promise((resolve, reject) =>{
         db.run(sql,(err) => {
             if (err) {
-                res.status(500).send(err)
-                return reject()
+                return reject(err)
             } else {
-                res.json({message:message})
                 return resolve()
             }
         })
@@ -77,42 +81,62 @@ const run = async (sql,db,res,message) => {
 
 // Create a new user
 app.post('/api/v1/users/',async(req, res) => {
-    //connect database
-    const db = new sqlite3.Database(dbPath)
+    if (!req.body.name || req.body.name ==="") {
+        res.status(400).send({error:"ユーザー名が指定されていません"})
+    } else {
+        //connect database
+        const db = new sqlite3.Database(dbPath)
 
-    const name = req.body.name
-    const profile = req.body.profile ? req.body.profile : ""
-    const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : ""
+        const name = req.body.name
+        const profile = req.body.profile ? req.body.profile : ""
+        const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : ""
 
-    await run(
-        `INSERT INTO users (name, profile, date_of_birth) VALUES ("${name}","${profile}","${dateOfBirth}")`,
-        db,
-        res,
-        "新規ユーザーを作成しました。")
-    db.close()
+        try {
+            //runは失敗したときにrejectを返しそうするとcatchがエラーとしてキャッチしてくれる。
+            await run(
+                `INSERT INTO users (name, profile, date_of_birth) VALUES ("${name}","${profile}","${dateOfBirth}")`,
+                db
+            )
+
+            res.status(201).send({message:"新規ユーザーを作成しました。"})
+        } catch (e) {
+            res.status(500).send({error:e})
+        }
+        db.close()
+    }
 })
 // Update user data
 app.put('/api/v1/users/:id',async(req, res) => {
-    //connect database
-    const db = new sqlite3.Database(dbPath)
-    const id = req.params.id
+    if (!req.body.name || req.body.name ==="") {
+        res.status(400).send({error:"ユーザー名が指定されていません"})
+    } else {
+        //connect database
+        const db = new sqlite3.Database(dbPath)
+        const id = req.params.id
 
-    //現在のユーザー情報を取得する
-    db.get(`SELECT * FROM users WHERE id=${id}`,async(err,row) => {
+        //現在のユーザー情報を取得する
+        db.get(`SELECT * FROM users WHERE id=${id}`,async(err,row) => {
+            if (!row) {
+                res.status(404).send({error:"指定されたユーザーが見つかりません"})
+            } else {
+                //resはjsonというメソッドを持っていて引数で指定したjsonをそのまま返す。
+                const name = req.body.name ? req.body.name : row.name
+                const profile = req.body.profile ? req.body.profile : row.profile
+                const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth
 
-        //resはjsonというメソッドを持っていて引数で指定したjsonをそのまま返す。
-        const name = req.body.name ? req.body.name : row.name
-        const profile = req.body.profile ? req.body.profile : row.profile
-        const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth
-
-        await run(
-            `UPDATE  users SET name="${name}",profile="${profile}",date_of_birth="${dateOfBirth}" WHERE id=${id}`,
-            db,
-            res,
-            "ユーザー情報を更新しました。"
-            )
-    })
-    db.close()
+                try {
+                    await run(
+                        `UPDATE  users SET name="${name}",profile="${profile}",date_of_birth="${dateOfBirth}" WHERE id=${id}`,
+                        db
+                    )
+                    res.status(200).send({message:"ユーザー情報を更新しました。"})
+                } catch (e) {
+                    res.status(500).send({error:e})
+                }
+            }
+        })
+        db.close()
+    }
 })
 
 // Delete user data
@@ -121,12 +145,23 @@ app.delete('/api/v1/users/:id',async(req, res) => {
     const db = new sqlite3.Database(dbPath)
     const id = req.params.id
 
-    await run(
-        `DELETE FROM users WHERE id=${id}`,
-        db,
-        res,
-        "ユーザー情報を削除しました。"
-        )
+    //現在のユーザー情報を取得する
+    db.get(`SELECT * FROM users WHERE id=${id}`,async(err,row) => {
+        if (!row) {
+            res.status(404).send({error:"指定されたユーザーが見つかりません"})
+        } else {
+            try {
+                await run(
+                    `DELETE FROM users WHERE id=${id}`,
+                    db
+                )
+                res.status(200).send({message: "ユーザーを削除しました。"})
+            } catch(e){
+                alert("aaa");
+                res.status(500).send({error:e})
+            }
+        }
+    })
     db.close()
 })
